Move response types to util/ folder
